backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
django-cors-headers==4.3.1
Pillow==10.2.0
pytest-mock==3.12.0
